+++ test/system/data_source/dec.py
@@ -4,12 +4,12 @@
 
 from __future__ import annotations
 
+import inspect
 import typing
 from functools import wraps
 from typing import Callable
 
 import pytest
-
 
 if typing.TYPE_CHECKING:
     from .fixtures.api.util import ApiFixture
@@ -43,11 +43,10 @@ def against(*fixtures: DataSourceFixture) -> Callable:
 
             fixture.before_test()
             ds_instance = fixture.get_ds_instance()
-            test_method(
-                self,
-                data_source=ds_instance,
-                ds_sleep=ds_sleep
-            )
+            kwargs = {'data_source': ds_instance, 'ds_sleep': ds_sleep}
+            if 'fixture_name' in inspect.signature(test_method).parameters:
+                kwargs['fixture_name'] = fixture.name
+            test_method(self, **kwargs)
             fixture.after_test()
 
         return wrapper
@@ -63,7 +62,7 @@ def against_api(*fixtures: ApiFixture) -> Callable:
     """
 
     params = [
-        pytest.param(dsf, dsf.url, dsf.sleep, id=dsf.name)
+        pytest.param(dsf, dsf.url, dsf.sleep, dsf.name, id=dsf.name)
         for dsf in fixtures
     ]
 
@@ -78,7 +77,7 @@ def against_api(*fixtures: ApiFixture) -> Callable:
             self,
             data_source: ApiFixture,
             url: str,
-            ds_sleep: Callable[[float], None]
+            ds_sleep: Callable[[float], None],
         ) -> None:
 
             # a little hacky, but the names must match
@@ -90,7 +89,7 @@ def against_api(*fixtures: ApiFixture) -> Callable:
                 self,
                 data_source=ds_instance,
                 url=url,
-                ds_sleep=ds_sleep
+                ds_sleep=ds_sleep,
             )
             fixture.after_test()