+++ test/system/data_source/test_cases/test_e2e.py
@@ -175,8 +175,8 @@ class TestEndToEnd:
         ]
 
         if data_source.write_mode['root'] == RelationWriteMode.SEPARATE:
-            data_source.upsert('related', related_objects)
-        data_source.upsert('root', data_objects)
+            data_source.upsert('related', related_objects, provenance='source1')
+        data_source.upsert('root', data_objects, provenance='source1')
 
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
@@ -310,13 +310,17 @@ class TestEndToEnd:
         assert len(seventeenth) == 1  # The archetype
 
     @against(*all_fixtures)
-    def test_multiple_upserts(self, data_source: OperableDataSource, ds_sleep):
+    def test_multiple_upserts(
+        self,
+        data_source: OperableDataSource,
+        ds_sleep
+    ):
         """
         Upsert a `DataObject` instance, and then upsert again to test upsert behaviour
         """
 
         obj1 = data_source.data_object_factory('root', '1', attributes={})
-        data_source.upsert('root', [obj1])
+        data_source.upsert('root', [obj1], provenance='source1')
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
         # they should all be present now
@@ -336,7 +340,7 @@ class TestEndToEnd:
             attributes={
                 'str_column': 'value2',
                 'int_column': 123,
-            },
+            }
         )
 
         obj1 = data_source.data_object_factory(
@@ -349,12 +353,12 @@ class TestEndToEnd:
                 'bool_column': False,
                 'list_column': ['item1', 'item2'],
             },
-            to_one={'related_object': rel1},
+            to_one={'related_object': rel1}
         )
 
         if data_source.write_mode['root'] == RelationWriteMode.SEPARATE:
-            data_source.upsert('related', [rel1])
-        returned_ = data_source.upsert('root', [obj1])
+            data_source.upsert('related', [rel1], provenance='source1')
+        returned_ = data_source.upsert('root', [obj1], provenance='source1')
 
         if data_source.return_mode['root'] == ReturnMode.POPULATED:
             returned_root = list(returned_)[0]
@@ -390,7 +394,7 @@ class TestEndToEnd:
                 'int_column': 456,
                 'bool_column': False,
                 'datetime_column': datetime(2024, 6, 6),
-            },
+            }
         )
 
         obj1 = data_source.data_object_factory(
@@ -402,12 +406,12 @@ class TestEndToEnd:
                 'bool_column': True,
                 'list_column': ['item1', 'item3'],
             },
-            to_one={'related_object': rel1},
+            to_one={'related_object': rel1}
         )
 
         if data_source.write_mode['root'] == RelationWriteMode.SEPARATE:
-            data_source.upsert('related', [rel1])
-        data_source.upsert('root', [obj1])
+            data_source.upsert('related', [rel1], provenance='source1')
+        data_source.upsert('root', [obj1], provenance='source1')
 
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
@@ -427,7 +431,7 @@ class TestEndToEnd:
         assert ret.list_column == ['item1', 'item2', 'item3']
 
         obj1 = data_source.data_object_factory('root', '1', attributes={})
-        data_source.upsert('root', [obj1])
+        data_source.upsert('root', [obj1], provenance='source1')
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
         fourth = list(data_source.get_by_ids('root', ['1']))
@@ -457,7 +461,7 @@ class TestEndToEnd:
             },
             to_one={'related_object': None},
         )
-        data_source.upsert('root', [obj1])
+        data_source.upsert('root', [obj1], provenance='source1')
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
         fifth = list(data_source.get_by_ids('root', ['1']))
@@ -469,6 +473,60 @@ class TestEndToEnd:
         assert ret.bool_column is None
         assert ret.related_object is None
         assert ret.list_column is None
+
+    @against(elastic, api_elastic)
+    def test_upsert_with_multiple_sources(self, data_source: OperableDataSource, ds_sleep):
+        """
+        Test that upserting with multiple sources works as expected, and provenance is correct.
+        """
+
+        rel_obj1 = data_source.data_object_factory(
+            'related',
+            '1',
+            attributes={'str_column': 'related_value'},
+        )
+        rel_obj2 = data_source.data_object_factory(
+            'related',
+            '2',
+            attributes={'str_column': 'related_value'},
+        )
+        root_obj = data_source.data_object_factory(
+            'root',
+            '1',
+            attributes={
+                'int_column': 1,
+                'str_column': '1'
+            },
+            to_one={
+                'related_object': rel_obj1
+            }
+        )
+        root_obj2 = data_source.data_object_factory(
+            'root',
+            '1',
+            attributes={
+                'int_column': 1,
+                'str_column': '1'
+            },
+            to_one={
+                'related_object': rel_obj2
+            }
+        )
+
+        if data_source.write_mode['root'] == RelationWriteMode.SEPARATE:
+            data_source.upsert('related', [rel_obj1], provenance='source1')
+            data_source.upsert('related', [rel_obj2], provenance='source2')
+        data_source.upsert('root', [root_obj], provenance='source3')
+        data_source.upsert('root', [root_obj2], provenance='source4')
+        ds_sleep(5)
+
+        first = list(data_source.get_by_ids('root', ['1']))
+        ret = first[0]
+        assert ret.int_column == 1
+        assert ret.str_column == '1'
+        assert ret.provenance['related_object']['source3'].id == '1'
+        assert ret.provenance['related_object']['source4'].id == '2'
+        assert ret.related_object.id == '1'
 
     @against(sql, api_sql)
     def test_upsert_dict_attributes(self, data_source: OperableDataSource, ds_sleep):
@@ -543,7 +601,7 @@ class TestEndToEnd:
             },
             to_one={'related_object': None},
         )
-        data_source.upsert('root', [obj1])
+        data_source.upsert('root', [obj1], provenance='source1')
         ds_sleep(2)  # Let Elastic settle down after the upsert
 
         fifth = list(data_source.get_by_ids('root', ['1']))
@@ -631,8 +689,8 @@ class TestEndToEnd:
                 'datetime_column': datetime(2024, 6, 6),
             },
         )
-        data_source.upsert('root', [obj1])
-        data_source.upsert('related', [rel1, rel2])
+        data_source.upsert('root', [obj1], provenance='source1')
+        data_source.upsert('related', [rel1, rel2], provenance='source1')
         ds_sleep(2)
         # they should all be present now
         first = list(data_source.get_by_ids('root', ['1']))
@@ -661,6 +719,7 @@ class TestEndToEnd:
             'root',
             [update],
             candidate_key=['str_column'],
+            provenance='source1',
         )
         ds_sleep(2)
         second = list(data_source.get_by_ids('root', ['1']))
@@ -689,7 +748,7 @@ class TestEndToEnd:
                 'list_column': ['item1', 'item3'],
             },
         )
-        data_source.update('root', [update], candidate_key=['int_column'])
+        data_source.update('root', [update], candidate_key=['int_column'], provenance='source1')
         ds_sleep(2)
         third = list(data_source.get_by_ids('root', ['1']))
         assert len(third) == 1
@@ -707,7 +766,7 @@ class TestEndToEnd:
         assert ret.list_column == ['item1', 'item2', 'item3']
 
         update = (None, {'int_column': 2})
-        data_source.update('root', [update], candidate_key=['int_column'])
+        data_source.update('root', [update], candidate_key=['int_column'], provenance='source1')
         ds_sleep(2)
         fourth = list(data_source.get_by_ids('root', ['1']))
         assert len(fourth) == 1
@@ -735,7 +794,7 @@ class TestEndToEnd:
                 'list_column': None,
             },
         )
-        data_source.update('root', [update], candidate_key=['int_column'])
+        data_source.update('root', [update], candidate_key=['int_column'], provenance='source1')
         ds_sleep(5)
         fifth = list(data_source.get_by_ids('root', ['1']))
         assert len(fifth) == 1
@@ -759,7 +818,7 @@ class TestEndToEnd:
             },
         )
         data_source.update(
-            'root', [update], candidate_key_func=lambda x: ['int_column']
+            'root', [update], candidate_key_func=lambda x: ['int_column'], provenance='source1'
         )
         ds_sleep(5)
         sixth = list(data_source.get_by_ids('root', ['1']))